+++ addons/docs/common/preset.js
@@ -26,20 +26,8 @@ function webpack(webpackConfig = {}, options = {}) {
       ...module,
       rules: [
         ...(module.rules || []),
-        // {
-        //   test: [/\.stories\.(jsx?$|ts?$)/],
-        //   enforce: 'pre',
-        //   use: [
-        //     {
-        //       loader: require.resolve('@storybook/addon-storysource/loader'),
-        //       options: {
-        //         injectParameters: true,
-        //       },
-        //     },
-        //   ],
-        // },
         {
-          test: /\.stories.mdx$/,
+          test: /\.(stories|story).mdx$/,
           use: [
             {
               loader: 'babel-loader',
@@ -55,7 +43,7 @@ function webpack(webpackConfig = {}, options = {}) {
         },
         {
           test: /\.mdx$/,
-          exclude: /\.stories.mdx$/,
+          exclude: /\.(stories|story).mdx$/,
           use: [
             {
               loader: 'babel-loader',
@@ -72,11 +60,7 @@ function webpack(webpackConfig = {}, options = {}) {
 }
 
 function addons(entry = []) {
-  return [
-    ...entry,
-    // require.resolve('@storybook/addon-storysource/register'),
-    require.resolve('../register'),
-  ];
+  return [...entry, require.resolve('../register')];
 }
 
 module.exports = { webpack, addons };