+++ app/react-native/src/preview/components/StoryView/index.tsx
@@ -5,15 +5,9 @@ import addons from '@storybook/addons';
 import Events from '@storybook/core-events';
 
 interface Props {
-  listenToEvents: boolean;
-  selection?: any;
-  storyFn?: any;
+  stories: any;
   url: string;
-}
-
-interface State {
-  storyFn?: any;
-  selection?: any;
+  onDevice?: boolean;
 }
 
 const HelpContainer = styled.View`
@@ -24,31 +18,21 @@ const HelpContainer = styled.View`
   justify-content: center;
 `;
 
-export default class StoryView extends Component<Props, State> {
+export default class StoryView extends Component<Props> {
   componentDidMount() {
-    if (this.props.listenToEvents) {
-      const channel = addons.getChannel();
-      channel.on(Events.SELECT_STORY, this.selectStory);
-      channel.on(Events.FORCE_RE_RENDER, this.forceReRender);
-    }
+    const channel = addons.getChannel();
+    channel.on(Events.STORY_RENDER, this.forceReRender);
+    channel.on(Events.FORCE_RE_RENDER, this.forceReRender);
   }
 
   componentWillUnmount() {
-    const { listenToEvents } = this.props;
-
-    if (listenToEvents) {
-      const channel = addons.getChannel();
-      channel.removeListener(Events.SELECT_STORY, this.selectStory);
-      channel.removeListener(Events.FORCE_RE_RENDER, this.forceReRender);
-    }
+    const channel = addons.getChannel();
+    channel.removeListener(Events.STORY_RENDER, this.forceReRender);
+    channel.removeListener(Events.FORCE_RE_RENDER, this.forceReRender);
   }
 
   forceReRender = () => {
     this.forceUpdate();
-  };
-
-  selectStory = (selection: any) => {
-    this.setState({ storyFn: selection.storyFn, selection });
   };
 
   renderHelp = () => {
@@ -75,41 +59,24 @@ export default class StoryView extends Component<Props, State> {
   );
 
   render() {
-    const { listenToEvents } = this.props;
+    const { onDevice, stories } = this.props;
 
-    if (listenToEvents) {
-      return this.renderListening();
-    } else {
-      return this.renderOnDevice();
+    const selection = stories.getSelection();
+
+    const { id, storyFn } = selection;
+
+    if (storyFn) {
+      return (
+        <View key={id} style={{ flex: 1 }}>
+          {storyFn()}
+        </View>
+      );
     }
+
+    if (onDevice) {
+      return this.renderOnDeviceUIHelp();
+    }
+
+    return this.renderHelp();
   }
-
-  renderListening = () => {
-    if (!this.state) {
-      return null;
-    }
-    const { storyFn, selection } = this.state;
-    const { kind, story } = selection;
-
-    return storyFn ? (
-      <View key={`${kind}:::${story}`} style={{ flex: 1 }}>
-        {storyFn()}
-      </View>
-    ) : (
-      this.renderHelp()
-    );
-  };
-
-  renderOnDevice = () => {
-    const { storyFn, selection } = this.props;
-    const { kind, story } = selection;
-
-    return storyFn ? (
-      <View key={`${kind}:::${story}`} style={{ flex: 1 }}>
-        {storyFn()}
-      </View>
-    ) : (
-      this.renderOnDeviceUIHelp()
-    );
-  };
 }